Self Check Issues TS, Not completed.
not sure what is going on here, but when you take an oppenent's piece
with your king

it replaces the king with the original piece of the same color that was
on the board.

I positioned the pieces so you can force a check wiht the white queen,
and test it.